requirement.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
google-api-core==2.5.0
google-auth==2.6.0
google-cloud-speech==2.13.0
googleapis-common-protos==1.55.0
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
numpy==1.22.2
pandas==1.4.0
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydub==0.25.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.8
